requirements.txt
Django==1.11.3
django-grappelli==2.10.1
django-nested-inline==0.3.7
pluggy==0.4.0
py==1.4.34
python-monkey-business==1.0.0
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
